==> 1_setup/functional_traits/1e_collect_adult_mobility.Rmd <==
---
title: 'Collect adult mobility traits'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/src/templates/ohara_hdr.html'
  pdf_document:
    toc: true
---

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)

library(oharac)
library(tidyverse)
library(here)
source(here('common_fxns.R'))
source('fb_slb_fxns.R')

```

# Summary

Gather data on mobility traits from vulnerability project (are there options on FishBase or SeaLifeBase?).  Gapfill using downfill (and upstream-downstream gapfill? only reasonable for continuous) methodology.

This paper (https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4548538/) seems to indicate that certain traits including mobility should be available in WoRMS...?

# Data

This script will use processed valid species traits from the species vulnerability project

* Butt, N. et al. 2021. 

# Methods

## Adult mobility: Read in data

Here we reassign mobility classifications to different bins, of ascending mobility.

* sessile < sedentary < mobile resident < migratory/nomadic

For species with multiple values, assign to highest mobility bin?  This can be done at the processing stage if necessary.

```{r}
mob_vals <- c('sessile'                  = 'sessile',
              'nearly sessile/sedentary' = 'sedentary',
              'passive'                  = 'sedentary',
              'mobile resident'          = 'mobile resident',
              'horizontal migrator'      = 'migratory/nomadic',
              'nomadic'                  = 'migratory/nomadic')
### note: vertical migrator gets ignored here; we're looking at horizontal range
              

mob_vuln <- get_vuln_traits() %>%
  filter(str_detect(trait, 'adult_mobility')) %>%
  group_by(taxon, spp_gp) %>%
  summarize(trait_value = paste0(trait_value, collapse = ';')) %>%
  ungroup() 

mob_vuln_downfill <- mob_vuln %>%
  downfill() 

mob_vuln_expand <- mob_vuln_downfill %>%
  mutate(trait_value = str_split(trait_value, ';')) %>%
  unnest(trait_value) %>%
  mutate(mob_cat = mob_vals[trait_value],
         mob_cat = factor(mob_cat, levels = unique(mob_vals), ordered = TRUE)) %>%
  filter(!is.na(mob_cat)) %>%
  distinct()

```

Can we upstream/downstream these values roughly, since they are an ordered factor?

```{r}
mob_vuln_gf <- mob_vuln_expand %>%
  full_join(assemble_worms()) %>%
  mutate(mob_num = as.numeric(mob_cat)) %>%
  gapfill_up_down('mob_num') %>%
  mutate(value = round(value)) %>% ### round to nearest bin
  mutate(mob_cat = levels(mob_vuln_expand$mob_cat)[value],
         value   = factor(mob_cat, levels = levels(mob_vuln_expand$mob_cat))) %>%
  filter(!is.na(value))
```


Write out the results.  Note the summary table contains potential duplicates where one species falls into multiple bins.

```{r}
table(mob_vuln_gf %>% select(class, mob_cat)) %>% knitr::kable()

mob_vuln_out <- mob_vuln_gf %>%
  select(species, trait, value, sd, nspp, gf_level) %>%
  # filter(gf_level <= 4) %>%
  distinct() %>%
  arrange(species, gf_level)
write_csv(mob_vuln_out, here('_data/traits_grouping/trait_mobility.csv'))
```
